feat: Add basic HTML
