+++ highscore/package.json
@@ -6,13 +6,16 @@
   "packageManager": "yarn@3.3.0",
   "dependencies": {
     "@google-cloud/storage": "^6.7.0",
+    "express": "^4.18.2",
     "google-spreadsheet": "^3.3.0",
     "robotjs": "^0.6.0",
     "typescript": "^4.9.3",
     "win-control": "^0.3.1"
   },
   "devDependencies": {
+    "@types/express": "^4.17.14",
     "@types/google-spreadsheet": "^3.3.0",
-    "@types/node": "^18.11.9"
+    "@types/node": "^18.11.9",
+    "body-parser": "^1.20.1"
   }
 }
