+++ highscore/src/Score.ts
@@ -1,13 +1,14 @@
-interface IScore {
+interface IPartialScore {
   score: number;
   name: string;
+}
+
+interface IScore extends IPartialScore {
   photograph: string;
   image: string;
   timestamp: number;
 }
 
-class Score {
-  constructor(public data: IScore) {};
-}
+const isValidScore = (input: IPartialScore | unknown): input is IPartialScore => !!input && typeof input === "object" && "score" in input && "name" in input
 
-export { IScore, Score };
+export { IScore, IPartialScore, isValidScore };
